Added support for shrinking
Two new options: SHRINK=D where D=0, 1, or 2, and SHRINK-VERBOSE.  The
default (SHRINK=0) removes impossible edges via strongly connected
components, in the usual way.

At SHRINK=1, the algorithm finds all REQUIRED edges (ie, those that
**MUST** be part of the solution), and then removes all other edges from
the same sender or to the same receiver as a required edge.  This is
slow, and is only worthwhile if you are using ITERATIONS.  The shrinking
takes a while, but makes the iterations much faster, so you can run more
iterations overall.

At SHRINK=2, it follows the previous approach by then finding and
removing all FORBIDDEN edges (ie, those that cannot be part of any
solution).  This part is fairly fast, but is limited by the speed of
first finding the REQUIRED EDGES.